+++ app/models/protonym.rb
@@ -36,7 +36,8 @@ class Protonym < TaxonName
     :validate_source_type,
     :new_parent_taxon_name,
     :name_is_latinized,
-    :name_is_valid_format
+    :name_is_valid_format,
+    :verbatim_author_without_digits
 
   after_create :create_otu, if: -> {self.also_create_otu}
 
@@ -866,6 +867,10 @@ class Protonym < TaxonName
 
   def name_is_latinized
     errors.add(:name, 'Name must be latinized, no digits or spaces allowed') if !is_latin?
+  end
+
+  def verbatim_author_without_digits
+    errors.add(:verbatim_author, 'Verbatim author may not contain digits, a year may be present') if verbatim_author =~ /\d/ 
   end
 
   def name_is_valid_format